requirements.txt
langchain-core
langchain-openai
neo4j
ord-schema
protobuf
tqdm
rdkit
python-dotenv
requests
pydantic
streamlit
pyvis
